+++ src/main/java/session11pageobjectinterfaces/examples/pageobjectinterface_maryam/Animal.java
@@ -1,6 +1,55 @@
 package session11pageobjectinterfaces.examples.pageobjectinterface_maryam;
+// top level animal class
+public class Animal {
 
-public abstract class Animal implements IAnimalInterface {
 
-}
+    interface IAnimal {
+        String getSound ();
+    }
+
+    public void activeAnimal(){
+
+        IAnimal dog = new Dog();
+        IAnimal cat = new Cat();
+        IAnimal cow = new Cow();
+
+
+        System.out.println("Dog Sound" + " " + dog.getSound());
+        System.out.println("Cat Sound" + " " + cat.getSound());
+        System.out.println("Cow Sound" + " " + cow.getSound());
+
+
+
+    }
+    private static class Dog implements IAnimal {
+
+        @Override
+        public String getSound() {
+            return "Woof";
+        }
+    }
+
+    private static class Cat implements IAnimal{
+        @Override
+        public String getSound() {
+            return "Meaow";
+        }
+
+        }
+    private static class Cow implements IAnimal {
+
+        @Override
+        public String getSound() {
+            return "Moo";
+        }
+
+    }
+
+    }
+
+
+
+
+
+
