--- src/main/java/session11pageobjectinterfaces/examples/pageobjectinterface_maryam/Cat.java
@@ -1,12 +1,0 @@
-package session11pageobjectinterfaces.examples.pageobjectinterface_maryam;
-
-public class Cat extends Animal{
-    @Override
-
-    public String getSound(){
-
-        return "Meaow";
-    }
-
-}
-
