--- src/main/java/session11pageobjectinterfaces/examples/pageobjectinterface_maryam/Dog.java
@@ -1,10 +1,0 @@
-package session11pageobjectinterfaces.examples.pageobjectinterface_maryam;
-
-public class Dog extends Animal{
-
-    @Override
-    public String getSound() {
-
-      return "Woof";
-    }
-}
