--- src/main/java/session11pageobjectinterfaces/examples/pageobjectinterface_maryam/IAnimalInterface.java
@@ -1,5 +1,0 @@
-package session11pageobjectinterfaces.examples.pageobjectinterface_maryam;
-
-public interface IAnimalInterface {
-    String getSound();
-}
