+++ src/main/java/session11pageobjectinterfaces/examples/pageobjectinterface_maryam/Main.java
@@ -3,12 +3,8 @@ package session11pageobjectinterfaces.examples.pageobjectinterface_maryam;
 public class Main {
     public static void main(String[] args) {
 
-        IAnimalInterface dog = new Dog();
-        IAnimalInterface cat = new Cat();
-
-        System.out.println("Dog Sound" + " " + dog.getSound());
-        System.out.println("Cat Sound" + " " + cat.getSound());
-
+       Animal animal = new Animal();
+       animal.activeAnimal();
 
 
     }
